decompose all the things
