+++ src/components/UniversalNavbarExpanded/Desktop/DropDownCta.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Spacer, Body, COLORS, TitleSmall } from '../index'
-import { DropdownLinkIcon } from '../UniversalNavbar/assets/icons.js'
-import { removeLastWord, getLastWord } from '../../helpers/splitLastWord'
+import { Spacer, Body, COLORS, TitleSmall } from '../../index'
+import { DropdownLinkIcon } from '../../UniversalNavbar/assets/icons.js'
+import { removeLastWord, getLastWord } from '../../../helpers/splitLastWord'
 import styles from './DropDownCta.module.scss'
 
 const DropDownCta = ({ title, subcopy }) => (
@@ -10,7 +10,7 @@ const DropDownCta = ({ title, subcopy }) => (
     <TitleSmall.Serif.Book500>
       <span>{removeLastWord(title)}</span>
       <div className={styles.inlineBlock}>
-        {getLastWord(title)}
+        <span>{getLastWord(title)}</span>
         <DropdownLinkIcon />
       </div>
     </TitleSmall.Serif.Book500>
