+++ src/components/UniversalNavbarExpanded/Mobile/AccordionNav.module.scss
@@ -1,0 +1,70 @@
+@import '../../Media/Media.scss';
+
+.accordion {
+  padding: var(--Space-40) 0;
+  margin-bottom: var(--Space-40);
+  border-bottom: solid 1px var(--White);
+
+  @include for-tablet-only {
+    padding-bottom: var(--Space-56);
+    margin-bottom: var(--Space-56);
+  }
+}
+
+.accordionItem {
+  &:not(:last-child) {
+    margin-bottom: var(--Space-24);
+
+    @include for-tablet-only {
+      margin-bottom: var(--Space-32);
+    }
+  }
+
+  &.active {
+    .accordionChildren {
+      display: block;
+    }
+
+    .accordionParent svg {
+      transform: scale(1, -1);
+    }
+  }
+
+  .accordionParent > div:first-child,
+  .accordionChild {
+    max-width: calc(100% - 14px);
+    padding-right: var(--Space-64);
+
+    @include for-tablet-only {
+      padding-right: 120px;
+    }
+  }
+}
+
+.accordionParent {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  cursor: pointer;
+
+  > svg {
+    width: var(--Space-12);
+    height: var(--Space-12);
+    margin-bottom: var(--Space-8);
+  }
+
+  &:focus {
+    text-decoration: underline;
+    outline: none;
+  }
+}
+
+.accordionChildren {
+  display: none;
+  padding-bottom: var(--Space-8);
+}
+
+.accordionChild {
+  padding-top: var(--Space-24);
+}
