+++ src/components/Variables.scss
@@ -5,3 +5,5 @@
 $z-index-navbar: 1;
 $z-index-modal: 2;
 $z-index-navbar-outer: 3;
+
+$navbar-border: solid 1px var(--GrayStrokeAndDisabled--translucent);
